Fixes issues with returning the entire site when adding large number of data sets by responding via a custom template when returning the Site. Adds longer timeout to http requests, give better error message for http request failures, keeps rolling when adding all data set when one does not download, fixes wording and data types on admin interface.

+++ grails-app/services/pmel/sdig/las/ErddapService.groovy
@@ -570,6 +570,10 @@ class ErddapService {
                     cancelFile.delete()
                     throw new Exception("Request canceled.")
                 }
+                if ( !temp_file.exists() ) {
+                    log.error("ERDDAP download failed.")
+                    throw new Exception("Unable to download ERDDAP data set.")
+                }
                 temp_file.renameTo(new File(netcdfFilename));
                 dt = new DateTime();
                 log.info("Tabledap tool renamed the netcdf file to " + netcdfFilename + " at " + dt.toString(ISODateTimeFormat.dateHourMinuteSecondMillis()));
